fix: resolve all Linux GCC build errors (locally validated with act)
CMakeLists.txt:
- Add qt6-base-private-dev and libxkbcommon-dev to Linux apt deps
  (required for Qt6::GuiPrivate private headers and XKB::XKB target)
- Fix qt6-charts package name: qt6-charts-dev -> libqt6charts6-dev
- Remove non-existent libqt6sql6-dev and libqt6printsupport6-dev
  (headers are included in qt6-base-dev and qt6-base-private-dev)
- Add QXlsx to SKIP_UNITY_BUILD_INCLUSION (unity build was merging
  Polymarket screen files causing fmt_vol/fmt_price redefinitions)
- Undefine Qt strict cast macros for QXlsx target only (GCC/Clang: -U,
  MSVC: /U) — QXlsx uses implicit const char* -> QString conversions

TradingTools.cpp:
- Cast int64_t fields to qint64 before inserting into QJsonObject
  (on Linux int64_t is 'long int', causing ambiguous QJsonValue conversion)

EquityChart.cpp:
- Cast c.timestamp to qint64 before multiply in std::min/std::max calls
  (mixed qint64/long int types fail template deduction on GCC)

All fixes validated locally via act (catthehacker/ubuntu:act-22.04).
Binary produced: build/FinceptTerminal (16MB). Only artifact upload
fails locally as expected (needs ACTIONS_RUNTIME_TOKEN from GitHub).

+++ fincept-qt/src/mcp/tools/TradingTools.cpp
@@ -290,9 +290,9 @@ std::vector<ToolDef> get_trading_tools() {
                 auto stats = trading::pt_get_stats(id);
                 return ToolResult::ok_data(QJsonObject{{"total_pnl", stats.total_pnl},
                                                        {"win_rate", stats.win_rate},
-                                                       {"total_trades", stats.total_trades},
-                                                       {"winning_trades", stats.winning_trades},
-                                                       {"losing_trades", stats.losing_trades},
+                                                       {"total_trades", static_cast<qint64>(stats.total_trades)},
+                                                       {"winning_trades", static_cast<qint64>(stats.winning_trades)},
+                                                       {"losing_trades", static_cast<qint64>(stats.losing_trades)},
                                                        {"largest_win", stats.largest_win},
                                                        {"largest_loss", stats.largest_loss}});
             } catch (const std::exception& e) {
